% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaas.R
\name{gaa_download}
\alias{gaa_download}
\title{Faz o download de um GAA a partir do link}
\usage{
gaa_download(link, path)
}
\arguments{
\item{link}{link do GAA}

\item{path}{pasta onde salvar o arquivo}
}
\value{
caminho do arquivo baixado
}
\description{
Faz o download de um GAA a partir do link
}
