% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaas.R
\name{gaa_instituicoes}
\alias{gaa_instituicoes}
\title{Instituicoes a partir do site do GAA}
\usage{
gaa_instituicoes(estado)
}
\arguments{
\item{estado}{sigla do estado}
}
\value{
tibble com as instituicoes
}
\description{
Instituicoes a partir do site do GAA
}
